made level to 30, as in sad, fixed a small code style error in physic.

+++ AMBroSIA/src/game/Physics.java
@@ -89,12 +89,13 @@ public class Physics implements Runnable {
 
         float[] velocity = gameObject.getVelocity();
         float[] acceleration = calculate2DAcceleration(gameObject.getHeading(), gameObject.getAcceleration());
-
+        
         int[] displacement = calculateDisplacement(velocity, acceleration, 1);
         //int[] displacement = calculateDisplacement(velocity, gameObject.getVelocity(), 1);
         gameObject.setX(gameObject.getX() + displacement[0]);
         gameObject.setY(gameObject.getY() + displacement[1]);
 
+       
         gameObject.setVelocity(calculateNewVelocity(gameObject, velocity, acceleration, 1));
         wrapAround(gameObject);
     }
@@ -241,7 +242,7 @@ public class Physics implements Runnable {
 
         acceleration2D[0] = (float) (acceleration * Math.cos(Math.toRadians(heading - 90)));
         acceleration2D[1] = (float) (acceleration * Math.sin(Math.toRadians(heading - 90)));
-
+        
         return acceleration2D;
     }
 
@@ -277,8 +278,8 @@ public class Physics implements Runnable {
 
 //        displacement[0] = Math.round((velocity[0] * time + 0.5f * acceleration[0] * (float) Math.pow(time, 2)));
 //        displacement[1] = Math.round((float) velocity[1] * time + 0.5f * acceleration[1] * (float) Math.pow(time, 2));
-        displacement[0] = Math.round((velocity[0] * time + 0.5f * acceleration[0]));
-        displacement[1] = Math.round((float) velocity[1] * time + 0.5f * acceleration[1]);
+        displacement[0] = Math.round(velocity[0] * time + 0.5f * acceleration[0]);
+        displacement[1] = Math.round(velocity[1] * time + 0.5f * acceleration[1]);
         
         return displacement;
     }
